+++ internal/handlers/nextDate.go
@@ -2,7 +2,7 @@ package handlers
 
 import (
 	"fmt"
-	"go_final_project/internal/utils"
+	"go_final_project/internal/models"
 	"net/http"
 	"time"
 )
@@ -16,23 +16,24 @@ import (
 //
 // Если входные параметры недействительны или при вычислении возникает ошибка, функция возвращает ответ HTTP 400 Bad Request.
 // В противном случае она устанавливает заголовок "Content-Type" в "text/plain" и выводит результат в формате "%s\n".
-func NextDate(w http.ResponseWriter, r *http.Request) {
+func (h *Handler) NextDate(w http.ResponseWriter, r *http.Request) {
 	now, err := time.Parse("20060102", r.FormValue("now"))
 	if err != nil {
-		http.Error(w, fmt.Sprintf("%s\nневерный формат now", err), http.StatusBadRequest)
+		h.SendErr(w, err, http.StatusBadRequest)
 		return
 	}
 	date := r.FormValue("date")
 	repeat := r.FormValue("repeat")
 	if date == "" || repeat == "" {
-		http.Error(w, fmt.Sprintf("%s\nневерный формат date или repeat", err), http.StatusBadRequest)
+		h.SendErr(w, err, http.StatusBadRequest)
 		return
 	}
-	next, err := utils.NextDate(now, date, repeat)
+	next, err := models.NextDate(now, date, repeat)
 	if err != nil {
-		http.Error(w, fmt.Sprintf("%s\n", err), http.StatusBadRequest)
+		h.SendErr(w, err, http.StatusBadRequest)
 		return
 	}
+	h.logger.Infof("next date updated")
 	w.Header().Set("Content-Type", "text/plain")
 	fmt.Fprintf(w, "%s\n", next)
 }
