--- internal/models/models.go
@@ -1,9 +1,0 @@
-package models
-
-type Task struct {
-	ID      string `json:"id"`
-	Title   string `json:"title"`
-	Date    string `json:"date"`
-	Repeat  string `json:"repeat"`
-	Comment string `json:"comment"`
-}
